submodules/callback_stop: Fix crash

+++ submodules/callback_stop.zsh
@@ -50,11 +50,14 @@ else
     fi
 fi
 
-user_data=($(ls -1 "${user_config}"))
-
-if [[ ${#user_data} -eq 0 ]]
+if [[ -d "${user_config}" ]]
 then
-    rmdir "${user_config}"
+    user_data=($(ls -1 "${user_config}"))
+
+    if [[ ${#user_data} -eq 0 ]]
+    then
+        rmdir "${user_config}"
+    fi
 fi
 
 . "${units}/stop.zsh"
